[Jyosmita]Added: Updated salary

+++ Model/SalaryModel/SalaryUpdateModel.cs
@@ -10,7 +10,7 @@ namespace EmployeeManagement_day27.Model.SalaryModel
 
         public int SalaryId { get; set; }
         public string Month { get; set; }
-        public int EmployeeSalary { get; set; }
+        public decimal EmployeeSalary { get; set; }
         public int EmployeeId { get; set; }
     }
